+++ src/App.tsx
@@ -1,7 +1,12 @@
 import { AppRouter } from '@/app/router/app-router';
+import { LocalizationProvider } from '@/features/localization/localization-provider';
 
 function App() {
-  return <AppRouter />;
+  return (
+    <LocalizationProvider>
+      <AppRouter />
+    </LocalizationProvider>
+  );
 }
 
 export default App;